Faststats algorithm for Update Statistics
Implementation of the new "faststats" algorithm, used in
conjunction with a Hive sample table created during bulk
load. The new algorithm makes use of Counting Bloom Filters
to determine the number of distinct values without
sorting. The use of a Hive sample table results in much
faster scan rates reading the data sampled during the
bulk load of a table.